+++ src/Promise/AbstractPromise.php
@@ -121,12 +121,12 @@ abstract class AbstractPromise implements PromiseInterface
     }
 
     /**
-     * @param array<int, mixed> $args
+     * @param array<int|string, mixed> $args
      */
     abstract protected function processToExecution(callable &$callable, array &$args): mixed;
 
     /**
-     * @param array<int, mixed> $args
+     * @param array<int|string, mixed> $args
      * @throws Throwable
      */
     private function call(?callable &$callable, array &$args): void